+++ CHANGELOG.rst
@@ -4,9 +4,18 @@ Changelog
 Version v0.13.1
 ---------------
 
+Improvements
+~~~~~~~~~~~~
+- Add python 3.9 tests.
+
+Bug Fixes
+~~~~~~~~~
+- Ensure that ids in frame reports are always np.int64 even when using libsonata 0.1.10.
+- Fix deprecation warnings.
+
 Removed
 ~~~~~~~
-- Drop python3.6 support
+- Drop python 3.6 support.
 
 
 Version v0.13.0